requirements.txt
django>=5.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0
PyPDF2>=3.0.1
python-docx>=1.1.2
requests>=2.31.0
python-whois>=0.9.4
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
dnspython>=2.6.1
gunicorn>=21.2.0
whitenoise>=6.6.0
gevent>=24.2.1
bleach>=6.1.0
django-ratelimit>=4.1.0
python-magic>=0.4.27
